requirements.txt
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.11.7
django-admin-sortable2==0.6.16
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.5.2
django-easy-select2==1.3.4
django-extensions==1.9.7
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==2.0.2
django-sekizai==0.10.0
django-taggit==0.22.2
django-treebeard==4.1.2
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-installer==0.9.7
djangocms-link==2.1.2
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.5
html5lib==0.9999999
olefile==0.44
Pillow==4.3.0
pydotplus==2.0.2
pyparsing==2.2.0
pytz==2018.3
six==1.11.0
typing==3.6.2
tzlocal==1.4
Unidecode==0.4.21